moved User model into its own app, refactored migrations to support
