+++ user/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
+from . import models
+
+
+@admin.register(models.User)
+class CustomUserAdmin(UserAdmin):
+    """Custom user admin."""
