+++ user/models.py
@@ -1,0 +1,20 @@
+import uuid
+from django.db import models
+from django.contrib.auth.models import AbstractUser
+
+
+class BaseModel(models.Model):
+    """Base model for the application. Uses UUID for pk."""
+    id = models.UUIDField(
+        primary_key=True,
+        editable=False,
+        default=uuid.uuid4,
+    )
+
+    class Meta:
+        """Metadata."""
+        abstract = True
+
+
+class User(BaseModel, AbstractUser):
+    """Custom user model."""
